+++ jrds-standalone/src/main/java/jrds/standalone/Jetty.java
@@ -43,7 +43,7 @@ public class Jetty extends CommandStarterImpl {
         logger.debug("Configuration: " + configuration);
 
         host = configuration.getProperty("jetty.host");
-        port = jrds.Util.parseStringNumber(configuration.getProperty("jetty.port"), port).intValue();
+        port = jrds.Util.parseStringNumber(configuration.getProperty("jetty.port"), port);
         propFileName = configuration.getProperty("propertiesFile", propFileName);
         webRoot = configuration.getProperty("webRoot", webRoot);
     }
